filter by type works
