+++ src/Hooks/usePokemon.js
@@ -1,15 +1,15 @@
-import { fetchPokemon } from '../services/fetch-utils';
+import { fetchPokemon, fetchPokeTypes } from '../services/fetch-utils';
 import { useState, useEffect } from 'react';
-
 
 export default function usePokemon() {
   const [pokemonList, setPokemonList] = useState([]);
-
+  const [type, setType] = useState('All');
+  const [typeList, setTypeList] = useState([]);
 
   useEffect(() => {
     const loadData = async () => {
       try {
-        const data = await fetchPokemon();
+        const data = await fetchPokemon(type);
         setPokemonList(data);
       } catch (e) {
     // eslint-disable-next-line no-console
@@ -18,6 +18,19 @@ export default function usePokemon() {
     };
   
     loadData();
-  }, []);
-  return { pokemonList };
+  }, [type]);
+  
+  useEffect(() => {
+    const getData = async () => {
+      try {
+        const data = await fetchPokeTypes();
+        setTypeList(data);
+      } catch (e) {
+        console.error(e);
+      }
+    };  
+    getData();
+
+  }, []); 
+  return { pokemonList, type, typeList, setType };
 }
