+++ src/services/fetch-utils.js
@@ -1,6 +1,10 @@
-export async function fetchPokemon() {
+export async function fetchPokemon(type) {
+  const params = new URLSearchParams();
+  if (type !== 'All') {
+    params.set('type', type);
+  }
   const resp = await fetch(
-    `https://pokedex-alchemy.herokuapp.com/api/pokedex?=perpage=150`
+    `https://pokedex-alchemy.herokuapp.com/api/pokedex?perPage=150&${params.toString()}`
   );
   const data = await resp.json();
   return data.results;
